% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{align_columns}
\alias{align_columns}
\title{Align matrix columns to reference}
\usage{
align_columns(A, B, verbose = FALSE)
}
\arguments{
\item{A}{Reference matrix with target column names}

\item{B}{Matrix to align}

\item{verbose}{Logical. Print alignment messages. Default FALSE.}
}
\value{
Matrix B with columns reordered to match A
}
\description{
Reorders and filters columns of matrix B to match column names of A.
}
\keyword{internal}
