% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_stan_backend}
\alias{check_stan_backend}
\title{Check availability of Stan backend}
\usage{
check_stan_backend(verbose = FALSE)
}
\arguments{
\item{verbose}{Logical. Print status messages. Default FALSE.}
}
\value{
Character string: "cmdstanr", "rstan", or "none"
}
\description{
Verifies if cmdstanr or rstan is available for model fitting.
}
\keyword{internal}
