% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_common_factor}
\alias{compute_common_factor}
\title{Compute common factor from matrix}
\usage{
compute_common_factor(Mz, T_train, use_train_loadings = FALSE, verbose = FALSE)
}
\arguments{
\item{Mz}{Standardized matrix (T x S)}

\item{T_train}{Number of training observations}

\item{use_train_loadings}{Logical. If TRUE, compute loadings only from
training data. Default FALSE.}

\item{verbose}{Logical. Print progress messages. Default FALSE.}
}
\value{
Numeric vector of factor scores (length T)
}
\description{
Extracts first principal component and standardizes using training period.
}
\keyword{internal}
