% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{print.ou_nested_mi}
\alias{print.ou_nested_mi}
\title{Print a multiple-imputation nested OU fit}
\usage{
\method{print}{ou_nested_mi}(x, ...)
}
\arguments{
\item{x}{An \code{ou_nested_mi} object from \code{\link{fit_ou_nested_mi}}.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a multiple-imputation nested OU fit
}
