% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{summarize_sv_sigmas}
\alias{summarize_sv_sigmas}
\title{Summarize stochastic volatility sigmas}
\usage{
summarize_sv_sigmas(fit)
}
\arguments{
\item{fit}{Fitted Stan model object}
}
\value{
List with component:
  \describe{
    \item{sigma_t}{Matrix (T x S) of median volatilities}
  }
}
\description{
Extracts median volatility paths from the SV component.
}
