% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{summary.ou_nested_mi}
\alias{summary.ou_nested_mi}
\alias{print.summary.ou_nested_mi}
\title{Summarize a multiple-imputation nested OU fit}
\usage{
\method{summary}{ou_nested_mi}(object, ...)

\method{print}{summary.ou_nested_mi}(x, ...)
}
\arguments{
\item{object}{An \code{ou_nested_mi} object from \code{\link{fit_ou_nested_mi}}.}

\item{...}{Ignored.}

\item{x}{A \code{summary.ou_nested_mi} object.}
}
\value{
An object of class \code{summary.ou_nested_mi}.
}
\description{
Summarize a multiple-imputation nested OU fit
}
