% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vmsg}
\alias{vmsg}
\title{Verbose message helper}
\usage{
vmsg(msg, verbose)
}
\arguments{
\item{msg}{Character. Message to print.}

\item{verbose}{Logical. Whether to print.}
}
\value{
NULL invisibly
}
\description{
Prints a message only if verbose is TRUE.
}
\keyword{internal}
