% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{build_beta_tmg_table}
\alias{build_beta_tmg_table}
\title{Build beta(TMG_t) table by sector and time}
\usage{
build_beta_tmg_table(fit, zTMG_use)
}
\arguments{
\item{fit}{Fitted Stan model object}

\item{zTMG_use}{Numeric vector. Standardized TMG series used in fitting.}
}
\value{
List with components:
  \describe{
    \item{beta_point}{Matrix (T x S) of beta values}
    \item{meta}{List with description metadata}
  }
}
\description{
Constructs the time-varying beta matrix using posterior medians.
}
